refactor(home): replace sidebar with single-column layout
