+++ src/app/_lib/cachedRecentPosts.ts
@@ -1,11 +1,11 @@
 import { getFilteredPosts } from '@/app/_lib/getFilteredPosts';
-import { getRecentAndUpdatedPosts } from '@/app/_lib/getRecentAndUpdatedPosts';
+import { getRecentPosts } from '@/app/_lib/getRecentPosts';
 
 /**
- * モジュールレベルでキャッシュされたデータ
+ * モジュールレベルでキャッシュされた最新記事データ
  *
- * IGNORE_TAGSで除外された記事を元に算出した最新記事と更新記事。
+ * IGNORE_TAGSで除外された記事を元に算出。
  * 全ページで共有され、計算は1回のみ実行される。
  */
 const filteredPosts = getFilteredPosts();
-export const { recentPosts, updatesPosts } = getRecentAndUpdatedPosts(filteredPosts);
+export const recentPosts = getRecentPosts(filteredPosts);
