+++ src/app/_lib/constants.ts
@@ -1,6 +1,5 @@
 export const POST_DISPLAY_LIMIT = 6;
 export const POPULAR_POST_DISPLAY_LIMIT = 6;
-export const UPDATED_POST_DISPLAY_LIMIT = 4;
 
 export const IGNORE_SLUGS: Set<string> = new Set(['20141105125846', '20171223014544']);
 export const IGNORE_TAGS: Set<string> = new Set(['名探偵コナン']);
